% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_white_test.R
\name{performWhiteTest}
\alias{performWhiteTest}
\title{White Test for Heteroscedasticity}
\usage{
performWhiteTest(model, data)
}
\arguments{
\item{model}{An object of class \code{lm}, representing a fitted linear model.}

\item{data}{A dataframe used to fit the \code{lm} model. It should include all variables used in the model.}
}
\value{
A list containing the test statistic, its p-value, and degrees of freedom.
}
\description{
Performs the White test for heteroscedasticity on a linear model object.
}
\examples{
data(mtcars)
model <- lm(mpg ~ wt + qsec, data = mtcars)
performWhiteTest(model, mtcars)
}
